some modifications

+++ calBmi/CalculBmi9.py
@@ -117,36 +117,40 @@ def readBmi():
 
 def buttdel():
     """
-    To earase last line 
-    of tensor.json
+    To earase last data if the usr would to delete it.
     """
-    try:
-        if os.path.getsize('./calBmi/doc_BMI9/file_bmi.json'):
-            with open('./calBmi/doc_BMI9/file_bmi.json', 'r') as file:
-                data = json.load(file)
-            for key, value in data.items():
-                print("Last value of bmi deleted")
-                print(value[-1])
-                del value[-1]
-            with open('./calBmi/doc_BMI9/file_bmi.json', 'w') as file:
-                data = json.dump(data, file, indent=4)
-            print("Last value of 'file_bmi.json' has been deleted !")
-    except FileNotFoundError:
-        print('+ Sorry, file asked not exist !')
+    MSB_War = messagebox.askyesno('Warning', '!!! Warning !!! If you' \
+        'continue, last result will be delete !!!')
+    if MSB_War == 1:
+        try:
+            if os.path.getsize('./calBmi/doc_BMI9/file_bmi.json'):
+                with open('./calBmi/doc_BMI9/file_bmi.json', 'r') as file:
+                    data = json.load(file)
+                for key, value in data.items():
+                    print("Last value of bmi deleted")
+                    print(value[-1])
+                    del value[-1]
+                with open('./calBmi/doc_BMI9/file_bmi.json', 'w') as file:
+                    data = json.dump(data, file, indent=4)
+                print("Last value of 'file_bmi.json' has been deleted !")
+        except FileNotFoundError:
+            print('+ Sorry, file asked not exist !')
 
-    try:
-        if os.path.getsize('./calBmi/doc_BMI9/file_kg.json'):
-            with open('./calBmi/doc_BMI9/file_kg.json', 'r') as file:
-                data = json.load(file)
-            for key, value in data.items():
-                print("Last value of weight deleted")
-                print(value[-1])
-                del value[-1]
-            with open('./calBmi/doc_BMI9/file_kg.json', 'w') as file:
-                data = json.dump(data, file, indent=4)
-            print("Last value of 'file_kg.json' has been deleted !")
-    except FileNotFoundError:
-        print('+ Sorry, file asked not exist !')
+        try:
+            if os.path.getsize('./calBmi/doc_BMI9/file_kg.json'):
+                with open('./calBmi/doc_BMI9/file_kg.json', 'r') as file:
+                    data = json.load(file)
+                for key, value in data.items():
+                    print("Last value of weight deleted")
+                    print(value[-1])
+                    del value[-1]
+                with open('./calBmi/doc_BMI9/file_kg.json', 'w') as file:
+                    data = json.dump(data, file, indent=4)
+                print("Last value of 'file_kg.json' has been deleted !")
+        except FileNotFoundError:
+            print('+ Sorry, file asked not exist !')
+    else:
+        MSBSHOW = messagebox.showinfo('Info', 'Ok, no data was deleted.')
 
 # To read name of patient for entry widget
 with open('./newpatient/entryfile9.txt', 'r') as filename:
@@ -216,39 +220,46 @@ textBox.grid(row=6, column=1, columnspan=2, pady=10)
 
 call_result = partial(call_result, textBox, number1, number2)
 
-buttonCal = Button(gui, text="1-Calcul", width=12, fg='yellow',
-    bg='turquoise4', activeforeground='gray40',
-    activebackground='turquoise2', command=call_result)
+buttonCal = Button(gui, text="1-Calcul", width=12, bd=3,
+    fg='yellow', bg='RoyalBlue3', activeforeground='gray40',
+    activebackground='turquoise2', highlightbackground='cyan',
+    command=call_result)
 buttonCal.grid(sticky='w', row=10, column=1)
 
-buttonCal = Button(gui, text="2-Save", width=12, fg='yellow',
-    bg='turquoise4', activeforeground='gray40',
-    activebackground='turquoise2', command=buttRecord)
-buttonCal.grid(sticky='w', row=11, column=1)
+buttonSave = Button(gui, text="2-Save", width=12, bd=3, 
+    fg='yellow', bg='RoyalBlue3', activeforeground='gray40',
+    activebackground='turquoise2', highlightbackground='cyan',
+    command=buttRecord)
+buttonSave.grid(sticky='w', row=11, column=1)
 
-buttonCal = Button(gui, text="Cancel last check", width=12, fg='yellow',
-    bg='coral', activeforeground='white',
-    activebackground='red', command=buttdel)
-buttonCal.grid(row=11, column=1, columnspan=2)
+buttonCancel = Button(gui, text="Cancel last check", width=12,
+    bd=3, fg='white', bg='coral', activeforeground='white',
+    activebackground='red', highlightbackground='orange',
+    command=buttdel)
+buttonCancel.grid(row=11, column=1, columnspan=2)
 
-buttonCal2 = Button(gui, text="Read", width=12, fg='light blue',
-    bg='turquoise4', activebackground='dark turquoise',
+buttonRead = Button(gui, text="Read", width=12, bd=3,
+    fg='coral', bg='RoyalBlue3',
+    activebackground='dark turquoise',
+    highlightbackground='cyan',
     command=readBmi)
-buttonCal2.grid(sticky='w', row=12, column=1)
+buttonRead.grid(sticky='w', row=12, column=1)
 
-buttonCal = Button(gui, text="Graph BMI", width=12, fg='blue2',
-    bg='turquoise3', activeforeground='gray30',
-    activebackground='turquoise', command=viewGraphicBmi)
-buttonCal.grid(sticky='e', row=10, column=2)
+buttonGraph1 = Button(gui, text="Graph BMI", width=12, bd=3,
+    fg='cyan', bg='RoyalBlue3', activeforeground='gray40',
+    activebackground='turquoise', highlightbackground='cyan',
+    command=viewGraphicBmi)
+buttonGraph1.grid(sticky='e', row=10, column=2, pady=10)
 
-buttonCal = Button(gui, text="Graph Weight", width=12, fg='blue2',
-    bg='turquoise3', activeforeground='gray30',
-    activebackground='turquoise', command=viewGraphicKilo)
-buttonCal.grid(sticky='e', row=11, column=2)
+buttonGraph2 = Button(gui, text="Graph Weight", width=12, bd=3,
+    fg='cyan', bg='RoyalBlue3', activeforeground='gray40',
+    activebackground='turquoise', highlightbackground='cyan',
+    command=viewGraphicKilo)
+buttonGraph2.grid(sticky='e', row=11, column=2)
 
-buttonCal2 = Button(gui, text="Quit", width=12, fg='cyan',
-    bg='gray30', activebackground='dark turquoise',
-    command=quit)
-buttonCal2.grid(sticky='e', row=12, column=2)
+buttonQuit = Button(gui, text="Quit", width=12, bd=3,
+    fg='white', bg='RoyalBlue3', activebackground='turquoise',
+    highlightbackground='cyan', command=quit)
+buttonQuit.grid(sticky='e', row=12, column=2, pady=10)
 
 gui.mainloop()
